Fix CR's issues
