+++ src/index.html
@@ -19,8 +19,7 @@
                 <div class="reload"><i class="fa fa-refresh" aria-hidden="true"></i></div>
             </header>
             <div class="task_input">
-                <input type="text" id="add_task" placeholder="Add to your list..."/>
-                <button type="button" class="button">&#9166;</button>          
+                <input type="text" id="add_task" placeholder="Add to your list..."/>          
             </div>
             <div class="task_container"></div>
             <div class="clear">
