+++ src/modules/add_task.js
@@ -1,0 +1,17 @@
+import Task, { addToLocalStorage } from './local_storage.js';
+
+import showActivity from './show_activity.js';
+
+const taskLists = JSON.parse(localStorage.getItem('data'));
+
+const addTask = () => {
+  const activity = document.querySelector('#add_task').value;
+  if (activity !== null) {
+    const tasks = new Task(activity, false, ((taskLists.length) + 1));
+    taskLists.push(tasks);
+    showActivity([tasks]);
+    addToLocalStorage(taskLists);
+  }
+};
+
+export default addTask;
