+++ src/styles.css
@@ -48,16 +48,6 @@ h1 {
   border-bottom: 1px solid #c4c6d8;
 }
 
-.task_input button {
-  border: none;
-  background-color: white;
-  align-self: center;
-  color: #c4c6d8;
-  margin-right: 1rem;
-  cursor: pointer;
-  font-size: 2rem;
-}
-
 input {
   font-size: 1.5rem;
   padding: 0.5rem;
